coordinator sets up and start calculation now
the coordinator sets up the calculation by calculating an initial GCD
and sending a product of this gcd to all gcd_clients. He then starts the
calculation by sending sendy to a selected few clients.
The coordinator is not handling any responses from the clients or a
kill/shutdown command or a reset.